Fix default access handling
The default access handling was written before custom fields were introduced. When custom fields were introduced the custom field import parser was written to consume all remaining values but the access handler was not updated to only write defaults for any unset values.

This PR updates the importer so the default access code doesn't look for further field, but, instead, only writes the default permissions which have not been overridden by the import.